movement2: Zoom camera out, use ScreenCamera
Since movement2 was built for 1080p, it looks bad at our tiny res. Zoom
out so you can see the height of the level as it was intended.

Switch to ScreenCamera since it's a much nicer presentation.

Simplify the camera code and make the two cameras both accept a node
path.