Make a scene transition effect
Based on https://www.gdquest.com/tutorial/godot/2d/scene-transition-rect/

Fade out before transitioning scenes so it's not so abrupt.

I tried to make it fade in by creating it automatically and by adding it
to the mainmenu and movement2/world scenes, but it didn't look good. In
the mainmenu, there was a flicker of the menu before the fade and
movement2 it didn't fade in at all.