Remove my funky tile offset
Use tex_offset on my tiles to make them match how tiles are placed in
the editor.

Now my cell positions all match the bottom corner of that cell.

Adjust player and goal snapped positions to be the centre of the cell.

Adjust player visual to line up with scene root y.