Rejigger the PauseMenu node to contain its children
The problem being that when the children weren't considered to be
inside the PauseMenu node (which was smaller than the PanelContainer)
so any signals to children of PanelContainer outside of PauseMenu were
lost. This manifested in the "Quit to menu" button closing the
PauseMenu Popup instead of quitting the game.