Setup Eggplant autoload to play music
Setup GameDef to have music in it that automatically plays when your
game starts. That way we can add music that plays while the game loads
and keeps playing without interruption during restarts.

Add music to my games.

Adjust volume of my music to match levels of Arkaruga and Triedge.